:recycle: イベントをmiddlewareとして扱うようにした
