+++ server/internal/cards/internal/events/ws/updated.go
@@ -8,79 +8,13 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/olahol/melody"
-	"github.com/traP-jp/h25s_15/internal/cards/internal/domain"
+	"github.com/traP-jp/h25s_15/internal/cards/internal/events"
 	"github.com/traP-jp/h25s_15/internal/games"
 )
 
-type EventCard struct {
-	ID    uuid.UUID `json:"id"`
-	Type  string    `json:"type"`
-	Value string    `json:"value"`
-}
-
-type CardUpdatedEvent struct {
-	Type              string      `json:"type"`
-	FieldCards        []EventCard `json:"fieldCards"`
-	Player0           []EventCard `json:"player0"`
-	Player0HandsLimit int         `json:"player0HandsLimit"`
-	Player1           []EventCard `json:"player1"`
-	Player1HandsLimit int         `json:"player1HandsLimit"`
-}
-
-func (e *Event) CardUpdated(ctx context.Context, gameID uuid.UUID) error {
-	cards, err := e.repo.GetActiveCards(ctx, gameID)
-	if err != nil {
-		return fmt.Errorf("get active cards: %w", err)
-	}
-
-	handLimits, err := e.repo.GetGameHandLimit(ctx, gameID)
-	if err != nil {
-		return fmt.Errorf("get game hand limit: %w", err)
-	}
-
-	fieldCards := make([]EventCard, 0, len(cards))
-	player0Cards := make([]EventCard, 0, len(cards))
-	player1Cards := make([]EventCard, 0, len(cards))
-
-	for _, card := range cards {
-		eventCard := EventCard{
-			ID:    card.ID,
-			Type:  string(card.Type),
-			Value: card.Value,
-		}
-
-		switch card.Location {
-		case domain.CardLocationField:
-			fieldCards = append(fieldCards, eventCard)
-		case domain.CardLocationHand:
-			if card.OwnerPlayerID == nil {
-				return fmt.Errorf("card in hand without owner player ID: %v", card)
-			}
-			ownerPlayerID := *card.OwnerPlayerID
-			switch ownerPlayerID {
-			case 0:
-				player0Cards = append(player0Cards, eventCard)
-			case 1:
-				player1Cards = append(player1Cards, eventCard)
-			default:
-				return fmt.Errorf("unexpected owner player ID: %d", ownerPlayerID)
-			}
-		default:
-			return fmt.Errorf("unexpected card location: %s", card.Location)
-		}
-	}
-
-	event := CardUpdatedEvent{
-		Type:              "cardUpdated",
-		FieldCards:        fieldCards,
-		Player0:           player0Cards,
-		Player0HandsLimit: handLimits[0],
-		Player1:           player1Cards,
-		Player1HandsLimit: handLimits[1],
-	}
-
+func (e *Event) CardUpdated(ctx context.Context, gameID uuid.UUID, event events.CardUpdatedEvent) error {
 	var eventJSON bytes.Buffer
-	err = json.NewEncoder(&eventJSON).Encode(event)
+	err := json.NewEncoder(&eventJSON).Encode(event)
 	if err != nil {
 		return fmt.Errorf("encode event: %w", err)
 	}
