+++ server/internal/cards/internal/events/ws/ws.go
@@ -2,17 +2,14 @@ package ws
 
 import (
 	"github.com/olahol/melody"
-	"github.com/traP-jp/h25s_15/internal/cards/internal/repository"
 )
 
 type Event struct {
 	m    *melody.Melody
-	repo repository.Repo
 }
 
-func New(m *melody.Melody, repo repository.Repo) *Event {
+func New(m *melody.Melody) *Event {
 	return &Event{
 		m:    m,
-		repo: repo,
 	}
 }
